create post added.

+++ src/blog/posts/views.py
@@ -38,8 +38,13 @@ def get_post(post_id: PrimaryKey):
 def create_post(post_in: PostCreate):
     """Create a new post."""
     # call service method create
-
-    pass
+    post = create(post_in)
+    if not post:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail=[{"msg": "Error happened in creating post."}],
+        )
+    return post
 
 
 @router.put("/{post_id}", response_model=PostRead)
@@ -59,12 +64,19 @@ def update_post(post_id: PrimaryKey, post_in: PostUpdate):
 @router.delete("/{post_id}", response_model=None)
 def delete_post(post_id: PrimaryKey):
     """Delete a post by ID."""
+    post = get(post_id=post_id)
+    if not post:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail=[{"msg": "Post with this id doesn't exist"}],
+        )
+    else:
 
-    # get post by id using service method
-    # if not post raise 404 HTTPException (check specification: https://fastapi.tiangolo.com/tutorial/handling-errors/ )
-    # else post exists, call delete from service
-    # usually no need to return anything from delete, sometimes the deleted id can be returned,
-    # but it's ok to return just HTTP 200 response
-    # (If there are no exceptions raised and no return, FastAPI automatically return HTTP 200)
+        # get post by id using service method
+        # if not post raise 404 HTTPException (check specification: https://fastapi.tiangolo.com/tutorial/handling-errors/ )
+        # else post exists, call delete from service
+        # usually no need to return anything from delete, sometimes the deleted id can be returned,
+        # but it's ok to return just HTTP 200 response
+        # (If there are no exceptions raised and no return, FastAPI automatically return HTTP 200)
 
-    pass
+        pass
